Move the Magento utility to its own file.
