+++ app/utils/utils.js
@@ -1,5 +1,4 @@
 import {createAction as actionCreator} from 'redux-act'
-import Immutable from 'immutable'
 
 // simplify redux-act createAction method.
 // usage: createAction('Update Campaign', 'id', 'update')
@@ -36,20 +35,4 @@ export const formEncode = (data) => {
         pairs.push(`${encodeURIComponent(k)}=${encodeURIComponent(data[k])}`)
     })
     return pairs.join('&').replace(/%20/g, '+')
-}
-
-/**
- * Extract all of the JSON pieces in 'text/x-magento-init' script
- * elements, and merge them together into a single configuration object
- *
- * Returns an Immutable Map ready for the Redux store.
- */
-export const extractMagentoJson = ($html) => {
-    return $html
-        .find('script[x-type="text/x-magento-init"]')
-        .map((_, item) => item.innerHTML)
-        .get()
-        .map(JSON.parse)
-        .map((item) => Immutable.fromJS(item))
-        .reduce((summary, item) => summary.mergeDeep(item), Immutable.Map())
 }
